refactor: qrcode and latex pipes
