+++ package.json
@@ -139,6 +139,7 @@
     "@types/marked": "^2.0.3",
     "@types/node": "^18.18.13",
     "@types/nouislider": "^15.0.0",
+    "@types/qrcode": "^1.5.5",
     "@types/swiper": "~4.2.0",
     "@typescript-eslint/eslint-plugin": "^6.13.1",
     "@typescript-eslint/parser": "^6.13.1",
