+++ src/app/shared/components/template/components/qr-code/qr-code.component.html
@@ -1,3 +1,3 @@
 <div class="image-wrapper">
-  <img src="{{ _row.value | qrCode | async }}" alt="{{ _row.value }}" />
+  <img src="{{ qrCodeData() | async }}" [alt]="_row.value" />
 </div>
