+++ src/app/shared/components/template/components/qr-code/qr-code.component.ts
@@ -1,4 +1,5 @@
-import { Component } from "@angular/core";
+import { Component, computed, effect } from "@angular/core";
+import QRCode from "qrcode";
 import { TemplateBaseComponent } from "../base";
 
 @Component({
@@ -6,4 +7,17 @@ import { TemplateBaseComponent } from "../base";
   templateUrl: "./qr-code.component.html",
   styleUrls: ["./qr-code.component.scss"],
 })
-export class TmplQRCodeComponent extends TemplateBaseComponent {}
+export class TmplQRCodeComponent extends TemplateBaseComponent {
+  /**
+   * Computed signal used to generate QR code from value input
+   *
+   * NOTE - as computed signals do not accept async values this method
+   * returns a promise which should be handled via async pipe
+   * */
+  public qrCodeData = computed((): Promise<string> => {
+    const value = this.value();
+    if (value && typeof value === "string") {
+      return QRCode.toDataURL(value);
+    }
+  });
+}
