+++ src/app/shared/components/template/pipes/index.ts
@@ -2,13 +2,11 @@ import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
 
 import { FilterDisplayComponentPipe } from "./filter-display-component.pipe";
-import { LatexPipe } from "./latex.pipe";
 import { MarkdownPipe } from "./markdown.pipe";
 import { NumberingPipe } from "./numbering.pipe";
 import { PLHAssetPipe } from "./plh-asset.pipe";
 import { StyleListPipe } from "./styleList.pipe";
 import { TranslatePipe } from "./translate.pipe";
-import { QRCodePipe } from "./qr-code.pipe";
 
 const TEMPLATE_PIPES = [
   FilterDisplayComponentPipe,
@@ -17,8 +15,6 @@ const TEMPLATE_PIPES = [
   NumberingPipe,
   StyleListPipe,
   TranslatePipe,
-  LatexPipe,
-  QRCodePipe,
 ];
 
 @NgModule({
